pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["xai_demo"]

[project]
name = "xai-demo"
version = "0.1.0"
description = "A demo application integrating xAI API"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "okpo", email = "okpox001@gmail.com"},
]
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "python-dotenv>=0.19.0",
    "streamlit>=1.24.0",
    "python-decouple>=3.8",
]

[project.urls]
Homepage = "https://github.com/Fpiol/ai--"
